Rename table.py->frame.py
